[worker] mcp split 1/6: extract proxy, runtime, instructions, federation helpers
