+++ src/mcp/mod.rs
@@ -31,7 +31,7 @@ use crate::db_discovery::{find_best_database, load_repos_config};
 use crate::embed::{EmbeddingService, ModelType};
 use crate::file::Language;
 use crate::fts::FtsStore;
-use crate::index::{IndexManager, SharedStores};
+use crate::index::SharedStores;
 use crate::rerank::{rrf_fusion, rrf_fusion_with_exact, vector_only, EXACT_MATCH_RRF_K};
 use crate::search::{adapt_rrf_k, boost_kind, detect_identifiers, detect_structural_intent};
 use crate::symbols::{SymbolIndexerRegistry, SymbolReference};
@@ -41,364 +41,26 @@ use regex::Regex;
 use rmcp::{
     handler::server::router::tool::ToolRouter,
     handler::server::wrapper::Parameters,
-    model::{
-        CallToolRequestParams, CallToolResult, Content, Implementation, ListToolsResult,
-        PaginatedRequestParams, ServerCapabilities, ServerInfo,
-    },
-    service::RequestContext,
-    tool, tool_handler, tool_router, ErrorData as McpError, RoleClient, RoleServer, ServerHandler,
+    model::{CallToolResult, Content, Implementation, ServerCapabilities, ServerInfo},
+    tool, tool_handler, tool_router, ErrorData as McpError, ServerHandler,
 };
 use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex};
-use tokio_util::sync::CancellationToken;
 
 // Re-export types
 pub use types::*;
 
-// ═══════════════════════════════════════════════════════════════════
-// MCP Proxy Service  (--mode client / --mode auto with serve detected)
-// ═══════════════════════════════════════════════════════════════════
+mod federation_helpers;
+mod instructions;
+mod proxy;
+mod runtime;
 
-/// Transparent stdio↔HTTP proxy with automatic reconnect.
-///
-/// When `codesearch mcp --mode client` is started by Claude Desktop:
-/// - Claude Desktop sends MCP requests over stdio
-/// - `McpProxyService` forwards every request to the running `codesearch serve` hub via HTTP
-/// - Responses flow back unchanged
-///
-/// This is the correct architecture for Claude Desktop: it has no repo context of its own
-/// and therefore cannot use `--mode local`. With `--mode client` it always connects to
-/// the serve hub, gaining access to all registered repos.
-///
-/// Only tool operations (`list_tools`, `call_tool`) are forwarded. Prompts, resources,
-/// and completion are not proxied — the serve hub does not expose them.
-///
-/// ## Reconnect
-///
-/// The peer is wrapped in `Arc<RwLock<Option<Peer>>>` so it can be hot-swapped when the
-/// serve connection drops and reconnects. During reconnection, tool calls return a
-/// descriptive "reconnecting" error so Claude Desktop can retry.
-///
-/// ## Idle disconnect / connect on demand
-///
-/// The peer is also `None` while the proxy is *deliberately* disconnected: after
-/// `CODESEARCH_MCP_PROXY_IDLE_DISCONNECT_SECS` (default
-/// `DEFAULT_MCP_PROXY_IDLE_DISCONNECT_SECS`, `0` disables) without a successful
-/// forwarded request, the idle-checker in `run_mcp_client` closes the HTTP MCP
-/// session so a scale-to-zero remote can suspend its replica. Every successful
-/// `list_tools` / `call_tool` stamps `last_activity`, which resets that window.
-///
-/// Because a closed session is indistinguishable from a dead one at the peer
-/// slot, `call_tool` / `list_tools` signal `connect_request_tx` on their first
-/// attempt whenever the slot is `None`, asking the main loop to connect *now*
-/// instead of waiting for the failure-path reconnect cadence. The existing
-/// bounded retry-with-backoff remains the fallback if that connect does not land
-/// within the retry budget.
-struct McpProxyService {
-    /// Shared peer handle — hot-swapped on reconnect.
-    /// `None` means we're reconnecting to serve; tool calls return a retry-able error.
-    peer: std::sync::Arc<tokio::sync::RwLock<Option<rmcp::service::Peer<RoleClient>>>>,
-    /// Signal to the main loop in `run_mcp_client` that the current peer is dead
-    /// and a fresh `connect_to_serve` should be attempted. Sent from `call_tool` /
-    /// `list_tools` when rmcp returns a transport-level error so we can recover
-    /// from server restarts and TCP keep-alive failures without bubbling the error
-    /// up to Claude Desktop.
-    disconnect_tx: tokio::sync::mpsc::Sender<()>,
-    /// Ask the main loop to run `connect_to_serve` immediately (capacity-1
-    /// channel — duplicate requests coalesce, "connect now" is idempotent).
-    /// Sent when a request arrives while the peer slot is empty.
-    connect_request_tx: tokio::sync::mpsc::Sender<()>,
-    /// When the last request was successfully forwarded to serve. Shared with the
-    /// idle-checker in `run_mcp_client`, which closes the connection once this is
-    /// older than the configured idle-disconnect window.
-    last_activity: Arc<Mutex<std::time::Instant>>,
-    /// Number of requests currently being forwarded. `last_activity` only advances
-    /// on completion, so without this a request that runs longer than the idle
-    /// window (a big search, a cold symbol rebuild) would have its own transport
-    /// closed underneath it. The idle-checker never disconnects while this is > 0.
-    in_flight: Arc<std::sync::atomic::AtomicUsize>,
-    /// Notified by the main loop's `connect_request_rx` arm whenever an on-demand
-    /// `connect_to_serve` attempt returns `Err` — i.e. serve refused the
-    /// connection outright, as opposed to still being slow to accept one. Lets
-    /// `await_peer` stop waiting immediately on a definitive failure instead of
-    /// polling out the rest of `PROXY_CONNECT_WAIT_MS` (previously ~20s per call
-    /// even when serve was known to be down within the first few milliseconds).
-    /// A slow-but-eventually-successful wake never touches this: it resolves by
-    /// the peer slot filling in, which `await_peer`'s own poll already catches.
-    connect_failed: Arc<tokio::sync::Notify>,
-}
-
-/// Keeps `McpProxyService::in_flight` incremented for its lifetime. A guard rather
-/// than paired add/sub calls because the forwarding loop has several early returns.
-struct InFlightGuard(Arc<std::sync::atomic::AtomicUsize>);
-
-impl InFlightGuard {
-    fn new(counter: &Arc<std::sync::atomic::AtomicUsize>) -> Self {
-        counter.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        Self(counter.clone())
-    }
-}
-
-impl Drop for InFlightGuard {
-    fn drop(&mut self) {
-        self.0.fetch_sub(1, std::sync::atomic::Ordering::SeqCst);
-    }
-}
-
-impl McpProxyService {
-    #[allow(dead_code)]
-    fn new(peer: rmcp::service::Peer<RoleClient>) -> Self {
-        // Direct constructor used by tests / single-shot scenarios.
-        // No reconnect plumbing — the dummy channels are never read.
-        let (tx, _rx) = tokio::sync::mpsc::channel(1);
-        let (connect_tx, _connect_rx) = tokio::sync::mpsc::channel(1);
-        Self {
-            peer: std::sync::Arc::new(tokio::sync::RwLock::new(Some(peer))),
-            disconnect_tx: tx,
-            connect_request_tx: connect_tx,
-            last_activity: Arc::new(Mutex::new(std::time::Instant::now())),
-            in_flight: Arc::new(std::sync::atomic::AtomicUsize::new(0)),
-            connect_failed: Arc::new(tokio::sync::Notify::new()),
-        }
-    }
-
-    /// Stamp "real traffic just flowed", resetting the idle-disconnect window.
-    fn mark_activity(&self) {
-        mark_proxy_activity(&self.last_activity);
-    }
-
-    /// Best-effort nudge to the main loop: connect to serve now. A full channel
-    /// already means "a connect is pending", a closed one means the loop is gone;
-    /// both are fine to ignore — the caller's own retry/backoff covers it.
-    fn request_connect(&self) {
-        let _ = self.connect_request_tx.try_send(());
-    }
-
-    /// Wait — bounded by `PROXY_CONNECT_WAIT_MS` — for the peer slot to be filled
-    /// after `request_connect`. Returns true as soon as a peer is available.
-    ///
-    /// Without this, a request arriving after an idle-close would burn its whole
-    /// retry budget (~1s) while the on-demand connect is still waking a
-    /// scaled-to-zero remote, and fail with "reconnecting" every single time.
-    async fn await_peer(&self) -> bool {
-        self.await_peer_bounded(PROXY_CONNECT_WAIT_MS).await
-    }
-
-    /// Core of `await_peer`, parameterized on the wait budget so it is unit
-    /// testable without actually waiting out `PROXY_CONNECT_WAIT_MS` (~20s).
-    /// Uses the production refusal-grace window; see
-    /// `await_peer_bounded_with_grace` for what that means and why it is its
-    /// own parameter.
-    async fn await_peer_bounded(&self, wait_ms: u64) -> bool {
-        self.await_peer_bounded_with_grace(wait_ms, CONNECT_REFUSAL_GRACE)
-            .await
-    }
-
-    /// Core of `await_peer_bounded`, additionally parameterized on the
-    /// refusal-grace window so *that* is unit testable without waiting out
-    /// `reconnect::INTERVAL_SECS` (~3s) for real.
-    ///
-    /// Polls the peer slot on `PROXY_RETRY_BACKOFF_MS` cadence, but also races
-    /// each poll against `connect_failed` so a definitive on-demand connect
-    /// failure (serve refused the connection, not merely slow to accept one)
-    /// clamps the remaining wait down to `refusal_grace` instead of polling
-    /// out the rest of `wait_ms`. A slow-but-still-in-progress wake never
-    /// fires `connect_failed` — it is only notified from an `Err` return of
-    /// `connect_to_serve` — so this does not shorten the legitimate
-    /// scale-to-zero wake path, only the case where serve is already known to
-    /// have refused this attempt.
-    ///
-    /// The clamp is deliberately *not* an immediate return: a refusal only
-    /// means this one on-demand attempt was refused, not that serve won't
-    /// recover — `run_mcp_client`'s own disconnect/reconnect cycle
-    /// (`reconnect::INTERVAL_SECS` later) can still land within the original
-    /// budget, e.g. when serve is mid-restart rather than genuinely down.
-    /// Returning immediately turned that case — previously transparent to the
-    /// caller, since the pre-fix full-budget poll caught the reconnect — into
-    /// a visible "reconnecting" error on the very first request after a
-    /// restart. Clamping to `refusal_grace` keeps most of the original fix's
-    /// win (a hard-down serve is still bounded well under the full `wait_ms`)
-    /// while still giving that recovery cycle room to land.
-    async fn await_peer_bounded_with_grace(
-        &self,
-        wait_ms: u64,
-        refusal_grace: std::time::Duration,
-    ) -> bool {
-        let mut deadline = std::time::Instant::now() + std::time::Duration::from_millis(wait_ms);
-        loop {
-            // Register for the next failure notification BEFORE checking the
-            // peer slot, so a failure landing between this check and the
-            // `select!` below cannot be missed (the standard tokio::sync::Notify
-            // idiom: create the `Notified` future first, await it second).
-            let failed = self.connect_failed.notified();
-            if self.peer.read().await.is_some() {
-                return true;
-            }
-            let now = std::time::Instant::now();
-            if now >= deadline {
-                return false;
-            }
-            let backoff = std::time::Duration::from_millis(PROXY_RETRY_BACKOFF_MS)
-                .min(deadline.saturating_duration_since(now));
-            tokio::select! {
-                _ = tokio::time::sleep(backoff) => {}
-                _ = failed => {
-                    // A concurrent successful connect could still have landed in
-                    // the instant before this notification; one last check keeps
-                    // that case correct instead of reporting a false failure.
-                    if self.peer.read().await.is_some() {
-                        return true;
-                    }
-                    deadline = deadline.min(now + refusal_grace);
-                }
-            }
-        }
-    }
-
-    /// On an empty peer slot (a deliberate idle-close or a real outage), ask the
-    /// main loop to connect *now* instead of waiting for the failure-path
-    /// reconnect cadence to notice, then give that connect a bounded window to
-    /// land. Returns true if a peer became available and the caller should
-    /// retry its forwarded call immediately.
-    ///
-    /// Only meaningful on the caller's first attempt (`attempt == 0`) — a
-    /// second empty slot means the on-demand connect already ran and fell
-    /// through to the ordinary retry/backoff path. Pulled out of `list_tools`/
-    /// `call_tool` because the two copies had already started to drift (see
-    /// review remarks on the commit that added this).
-    async fn try_on_demand_connect(&self) -> bool {
-        self.request_connect();
-        self.await_peer().await
-    }
-
-    /// Force a reconnect: clear the shared peer and signal the main loop in
-    /// `run_mcp_client` to call `connect_to_serve` again. Brief sleep gives
-    /// the main loop time to actually reconnect before the caller retries.
-    async fn force_reconnect(&self) {
-        *self.peer.write().await = None;
-        let _ = self.disconnect_tx.send(()).await;
-        tokio::time::sleep(std::time::Duration::from_millis(
-            crate::mcp::PROXY_RETRY_BACKOFF_MS,
-        ))
-        .await;
-    }
-}
-
-/// Maximum number of attempts when forwarding a request to serve.
-/// Each retry includes a forced reconnect, so this also bounds reconnect attempts
-/// per individual tool call.
-const PROXY_MAX_RETRY_ATTEMPTS: u32 = 3;
-
-/// Backoff between proxy retries, also used as the post-reconnect settle delay.
-const PROXY_RETRY_BACKOFF_MS: u64 = 500;
-
-/// How long a request may wait for an on-demand connect (after an idle-close, or
-/// while serve is still starting) before falling back to the retry/backoff path.
-///
-/// Sized for a scale-to-zero host: the remote's ingress *holds* the request while
-/// it activates a suspended replica, so the connect itself can legitimately take
-/// several seconds. Waiting here is strictly better than returning "reconnecting"
-/// on the first call after every idle period.
-const PROXY_CONNECT_WAIT_MS: u64 = 20_000;
-
-/// How long `await_peer_bounded` still waits after a definitive on-demand
-/// connect refusal, instead of returning immediately or polling out the rest
-/// of `PROXY_CONNECT_WAIT_MS`.
-///
-/// Sized to cover `run_mcp_client`'s own disconnect/reconnect cycle
-/// (`reconnect::INTERVAL_SECS`, ~3s) plus margin for the ~100ms synthetic-
-/// disconnect delay and `connect_to_serve`'s own latency — so a serve that is
-/// merely mid-restart still recovers transparently within this window,
-/// exactly as it did before the refusal short-circuit existed, while a
-/// genuinely-down serve is still bounded well under the full ~20s budget.
-const CONNECT_REFUSAL_GRACE: std::time::Duration =
-    std::time::Duration::from_millis(reconnect::INTERVAL_SECS * 1_000 + 1_000);
-
-/// Record a definitive on-demand connect refusal: wake any `await_peer_bounded`
-/// callers immediately (via `connect_failed`) instead of leaving them to poll
-/// out their full budget for a refusal that is already known, then seed a
-/// synthetic disconnect so `run_mcp_client`'s own disconnect/reconnect cycle
-/// picks it up. A genuinely slow wake never reaches this function — it
-/// resolves via the `Ok` branch in the caller once the peer slot fills in —
-/// so this does not shorten a legitimate scale-to-zero wake, only a refusal.
-///
-/// Pulled out of `run_mcp_client`'s `connect_request_rx` arm so the one line
-/// that makes `await_peer_bounded`'s refusal short-circuit real in production
-/// is covered by a test that calls this function directly, not only by tests
-/// that call `connect_failed.notify_waiters()` themselves in isolation —
-/// those pin how `await_peer_bounded` *reacts* to a notification, but nothing
-/// previously pinned that this call site still *fires* one: deleting this
-/// function's body left the full suite green.
-fn note_connect_failure(
-    connect_failed: &tokio::sync::Notify,
-    disconnect_tx: &tokio::sync::mpsc::Sender<()>,
-) {
-    connect_failed.notify_waiters();
-    let tx = disconnect_tx.clone();
-    tokio::spawn(async move {
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        let _ = tx.send(()).await;
-    });
-}
-
-/// Heuristic: does this error message describe a transport-level failure
-/// (broken TCP, server gone, stale keep-alive, stale session) that warrants
-/// a forced reconnect + retry, as opposed to a real tool-level error that
-/// the caller should see?
-fn is_transport_error_msg(msg: &str) -> bool {
-    msg.contains("Transport send error")
-        || msg.contains("error sending request")
-        || msg.contains("Transport error")
-        || msg.contains("connection closed")
-        || msg.contains("error decoding response body")
-        || msg.contains("Session not found")
-        || msg.contains("404")
-}
-
-/// Reconnect-related constants for the MCP proxy.
-mod reconnect {
-    /// How long to wait between reconnect attempts.
-    pub const INTERVAL_SECS: u64 = 3;
-    /// Maximum total time to spend trying to reconnect before giving up.
-    pub const MAX_DURATION_SECS: u64 = 300; // 5 minutes
-}
-
-/// Record the current instant as the proxy's most recent activity.
-fn mark_proxy_activity(last_activity: &Arc<Mutex<std::time::Instant>>) {
-    if let Ok(mut slot) = last_activity.lock() {
-        *slot = std::time::Instant::now();
-    }
-}
-
-/// Resolve the MCP proxy idle-disconnect window: explicit value → env var →
-/// `DEFAULT_MCP_PROXY_IDLE_DISCONNECT_SECS`. Mirrors how `run_serve` resolves its
-/// own `idle_suspend_secs`. `0` means "never idle-disconnect".
-fn resolve_proxy_idle_disconnect_secs(explicit: Option<u64>) -> u64 {
-    explicit
-        .or_else(|| {
-            std::env::var(crate::constants::MCP_PROXY_IDLE_DISCONNECT_SECS_ENV)
-                .ok()
-                .and_then(|s| s.trim().parse().ok())
-        })
-        .unwrap_or(crate::constants::DEFAULT_MCP_PROXY_IDLE_DISCONNECT_SECS)
-}
-
-/// Has the proxy been idle long enough to close its connection to serve?
-///
-/// `threshold_secs == 0` disables idle-disconnect, so this always returns false.
-/// `now` is a parameter (rather than read from the clock) purely so this is unit
-/// testable without sleeping.
-fn is_idle(
-    last_activity: std::time::Instant,
-    threshold_secs: u64,
-    now: std::time::Instant,
-) -> bool {
-    if threshold_secs == 0 {
-        return false;
-    }
-    now.saturating_duration_since(last_activity).as_secs() >= threshold_secs
-}
+// Re-export the extracted modules' items so every existing path keeps working:
+// `super::X` from sibling test files and `crate::mcp::X` from serve/cli.
+pub(crate) use federation_helpers::*;
+pub(crate) use instructions::*;
+pub use runtime::*;
 
 /// Resolve the `find_impact` wall-clock budget: env var →
 /// `DEFAULT_FIND_IMPACT_BUDGET_SECS`. `0` disables the budget (unbounded
@@ -455,20 +117,6 @@ where
     }
 }
 
-#[cfg(test)]
-#[path = "proxy_idle_tests.rs"]
-mod proxy_idle_tests;
-
-/// Unit tests for `await_peer_bounded`'s refusal clamp and the
-/// `note_connect_failure` call site that fires it in production, isolated
-/// from the full `run_mcp_client` loop by parameterizing the wait budget (and,
-/// for the clamp itself, the refusal-grace window) so these run in
-/// milliseconds instead of the real `PROXY_CONNECT_WAIT_MS` (~20s) or
-/// `reconnect::INTERVAL_SECS` (~3s).
-#[cfg(test)]
-#[path = "await_peer_tests.rs"]
-mod await_peer_tests;
-
 /// Tests for the `find_impact` wall-clock budget: sleeping-handler race
 /// tests (generic lookup future) plus `#[serial]` env-resolution tests.
 #[cfg(test)]
@@ -479,140 +127,6 @@ mod find_impact_tests;
 /// lookups: in-flight tracking keyed by (project, symbol) so a retry
 /// observes progress or the warm result instead of restarting cold.
 mod find_impact_tracker;
-
-impl ServerHandler for McpProxyService {
-    fn get_info(&self) -> ServerInfo {
-        ServerInfo::new(ServerCapabilities::builder().enable_tools().build())
-            .with_server_info(
-                Implementation::new("codesearch", env!("CARGO_PKG_VERSION"))
-                    .with_title("codesearch (serve proxy)"),
-            )
-            .with_instructions(
-                "Proxy to a running codesearch serve hub. All tool calls are forwarded to the hub.",
-            )
-    }
-
-    async fn list_tools(
-        &self,
-        request: Option<PaginatedRequestParams>,
-        _cx: RequestContext<RoleServer>,
-    ) -> Result<ListToolsResult, McpError> {
-        let _in_flight = InFlightGuard::new(&self.in_flight);
-        let mut last_err: Option<String> = None;
-        for attempt in 0..PROXY_MAX_RETRY_ATTEMPTS {
-            let peer = self.peer.read().await.clone();
-            match peer {
-                Some(p) => match p.list_tools(request.clone()).await {
-                    Ok(r) => {
-                        self.mark_activity();
-                        return Ok(r);
-                    }
-                    Err(e) => {
-                        let msg = e.to_string();
-                        if !is_transport_error_msg(&msg) || attempt >= PROXY_MAX_RETRY_ATTEMPTS - 1
-                        {
-                            return Err(McpError::internal_error(msg, None));
-                        }
-                        tracing::warn!(
-                            "list_tools attempt {}/{} failed (transport): {} — forcing reconnect",
-                            attempt + 1,
-                            PROXY_MAX_RETRY_ATTEMPTS,
-                            msg
-                        );
-                        last_err = Some(msg);
-                        self.force_reconnect().await;
-                    }
-                },
-                None => {
-                    // Empty peer slot: either a deliberate idle-close or a real
-                    // outage. `try_on_demand_connect` asks the main loop to
-                    // connect *now* rather than waiting for the failure-path
-                    // reconnect cadence to notice, bounded so we still fall back
-                    // to the ordinary retry/backoff below if it doesn't land.
-                    if attempt == 0 && self.try_on_demand_connect().await {
-                        continue;
-                    }
-                    if attempt < PROXY_MAX_RETRY_ATTEMPTS - 1 {
-                        tokio::time::sleep(std::time::Duration::from_millis(
-                            PROXY_RETRY_BACKOFF_MS,
-                        ))
-                        .await;
-                        continue;
-                    }
-                    return Err(McpError::internal_error(
-                        "codesearch serve is reconnecting — please retry in a moment".to_string(),
-                        None,
-                    ));
-                }
-            }
-        }
-        Err(McpError::internal_error(
-            last_err.unwrap_or_else(|| "transport error after retries".to_string()),
-            None,
-        ))
-    }
-
-    async fn call_tool(
-        &self,
-        request: CallToolRequestParams,
-        _cx: RequestContext<RoleServer>,
-    ) -> Result<CallToolResult, McpError> {
-        let _in_flight = InFlightGuard::new(&self.in_flight);
-        let mut last_err: Option<String> = None;
-        for attempt in 0..PROXY_MAX_RETRY_ATTEMPTS {
-            let peer = self.peer.read().await.clone();
-            match peer {
-                Some(p) => match p.call_tool(request.clone()).await {
-                    Ok(r) => {
-                        self.mark_activity();
-                        return Ok(r);
-                    }
-                    Err(e) => {
-                        let msg = e.to_string();
-                        if !is_transport_error_msg(&msg) || attempt >= PROXY_MAX_RETRY_ATTEMPTS - 1
-                        {
-                            return Err(McpError::internal_error(msg, None));
-                        }
-                        tracing::warn!(
-                            "call_tool('{}') attempt {}/{} failed (transport): {} — forcing reconnect",
-                            request.name,
-                            attempt + 1,
-                            PROXY_MAX_RETRY_ATTEMPTS,
-                            msg
-                        );
-                        last_err = Some(msg);
-                        self.force_reconnect().await;
-                    }
-                },
-                None => {
-                    // Empty peer slot: either a deliberate idle-close or a real
-                    // outage. `try_on_demand_connect` asks the main loop to
-                    // connect *now* rather than waiting for the failure-path
-                    // reconnect cadence to notice, bounded so we still fall back
-                    // to the ordinary retry/backoff below if it doesn't land.
-                    if attempt == 0 && self.try_on_demand_connect().await {
-                        continue;
-                    }
-                    if attempt < PROXY_MAX_RETRY_ATTEMPTS - 1 {
-                        tokio::time::sleep(std::time::Duration::from_millis(
-                            PROXY_RETRY_BACKOFF_MS,
-                        ))
-                        .await;
-                        continue;
-                    }
-                    return Err(McpError::internal_error(
-                        "codesearch serve is reconnecting — please retry in a moment".to_string(),
-                        None,
-                    ));
-                }
-            }
-        }
-        Err(McpError::internal_error(
-            last_err.unwrap_or_else(|| "transport error after retries".to_string()),
-            None,
-        ))
-    }
-}
 
 /// Read model short-name and dimensions from a database's `metadata.json`.
 /// Returns `(model_name, dimensions)`, defaulting to `("unknown", DEFAULT_EMBEDDING_DIMENSIONS)`.
@@ -7063,264 +6577,6 @@ fn contains_symbol_as_word(sig: &str, symbol: &str) -> bool {
     false
 }
 
-/// MCP server instructions template (pre-substitution). Kept as a named const so
-/// the line-count and deprecated-alias tests can validate it directly without
-/// fragile `include_str!` source-text searching or instantiating the service.
-/// Substitution uses `str::replace` (not `format!`) because `format!` requires a
-/// literal; the placeholders are unique tokens that don't appear in the prose.
-/// See `test_instructions_max_50_lines` / `test_no_deprecated_tool_aliases`.
-const INSTRUCTIONS_TEMPLATE: &str = r#"codesearch — semantic code search + symbol impact analysis.
-
-WHEN TO USE codesearch (prefer over grep/glob):
-  Good for: semantic or cross-file lookup, unknown file paths, symbol navigation,
-    "where is X implemented", "find usages of Y", "how does Z flow through the code"
-  Not for: a single known file (just read it), trivial one-line edits,
-    exact literal patterns where plain grep is faster
-
-SERVICE-MODE NOTES (codesearch serve, esp. on another host):
-  - Paths come from the SERVER's filesystem. Use get_chunk to read content;
-    don't try to open returned paths locally.
-  - Not every directory is indexed (e.g. .venv, node_modules, build/). If a
-    search returns nothing, the dir may be unindexed — ask, don't grep blindly.
-
-PICK THE RIGHT TOOL FOR THE TASK:
-  "who calls X?" / "what breaks if I rename X?"
-    → find_impact (precise SCIP call-graph; if no backend for the language, it says so → then use find kind="usages")
-  "find code about X" / "how does X work" / "show me X"
-    → search(mode="semantic") — concepts + synonyms + identifiers
-  exact syntax like Vec<T> / foo = null / a::b
-    → search(mode="literal", regex=true) — patterns semantic can't match
-  "where is X defined?" / "what does file X import?"
-    → find(kind="definition" | "imports")
-  "show all symbols in file X" / "code like chunk Y"
-    → explore(kind="outline" | "similar")
-  read chunk content → get_chunk(chunk_id)
-  index health / repo list → status
-
-RULES:
-  - search(semantic) is the DEFAULT for code lookup. Don't skip it.
-  - For "who calls X" / impact analysis, try find_impact first; fall back to find(kind="usages") only if find_impact reports no backend.
-  - NEVER use literal as first search unless you need exact syntax.
-  - project or group is REQUIRED in multi-repo mode.
-
-Mode: {mode}
-Project: {project}
-Database: {db} ({exists})
-Model: {model} ({dims}d)
-"#;
-
-// ════════════════════════════════════════════════════════════════
-// Federation helpers (module-scope) — merge / parse / convert.
-// ════════════════════════════════════════════════════════════════
-
-/// RRF-interleave several disjoint ranked lists into one ranked list.
-///
-/// Each list is assumed already ranked best-first and disjoint from the others
-/// (local repos vs. distinct remote peers). An item's merged score is
-/// `1/(k + rank_in_own_list + 1)` (classic Reciprocal Rank Fusion with a `+1`
-/// so the top hit never exceeds `1/k`). The union is sorted by score desc with a
-/// stable source-order tiebreak, then truncated to `limit`.
-fn merge_ranked_lists(
-    lists: Vec<Vec<SearchResultItem>>,
-    k: f32,
-    limit: usize,
-) -> Vec<SearchResultItem> {
-    let mut merged: Vec<(f32, usize, SearchResultItem)> = Vec::new();
-    let mut order = 0usize;
-    for list in lists {
-        for (rank, item) in list.into_iter().enumerate() {
-            let score = 1.0 / (k + rank as f32 + 1.0);
-            merged.push((score, order, item));
-            order += 1;
-        }
-    }
-    // Sort by score desc; tiebreak on insertion order for stable, predictable
-    // output (local list first, then remotes in config order).
-    merged.sort_by(|a, b| {
-        b.0.partial_cmp(&a.0)
-            .unwrap_or(std::cmp::Ordering::Equal)
-            .then(a.1.cmp(&b.1))
-    });
-    merged
-        .into_iter()
-        .take(limit)
-        .map(|(score, _, mut it)| {
-            it.score = score;
-            it
-        })
-        .collect()
-}
-
-/// Extract the rendered tool payload from a `CallToolResult` and re-parse it as
-/// local `SearchResultItem`s. Works for both semantic and literal modes — the
-/// rendered JSON always has a top-level `results` array.
-fn parse_search_items_from_call_result(
-    result: &CallToolResult,
-    mode: &str,
-) -> Vec<SearchResultItem> {
-    let text = extract_call_tool_text(result);
-    let value: serde_json::Value = match serde_json::from_str(&text) {
-        Ok(v) => v,
-        Err(_) => return Vec::new(),
-    };
-    let results = match value.get("results").and_then(|r| r.as_array()) {
-        Some(arr) => arr,
-        None => return Vec::new(),
-    };
-    match mode {
-        "semantic" => results
-            .iter()
-            .filter_map(|v| serde_json::from_value::<SearchResultItem>(v.clone()).ok())
-            .collect(),
-        // Literal items lack `chunk_id`; map their `snippet` into `content` so
-        // the merged list renders uniformly. The absent id stays `None` —
-        // fabricating `0` here once let callers build a bogus
-        // `<peer>/<alias>:0` chunk_ref that `get_chunk` silently resolved to
-        // an unrelated chunk.
-        _ => results
-            .iter()
-            .map(|v| SearchResultItem {
-                chunk_id: v.get("chunk_id").and_then(|c| c.as_u64()).map(|c| c as u32),
-                path: v
-                    .get("path")
-                    .and_then(|p| p.as_str())
-                    .unwrap_or("")
-                    .to_string(),
-                start_line: v.get("start_line").and_then(|n| n.as_u64()).unwrap_or(0) as usize,
-                end_line: v.get("end_line").and_then(|n| n.as_u64()).unwrap_or(0) as usize,
-                kind: v
-                    .get("kind")
-                    .and_then(|k| k.as_str())
-                    .unwrap_or("")
-                    .to_string(),
-                score: v.get("score").and_then(|s| s.as_f64()).unwrap_or(0.0) as f32,
-                signature: v
-                    .get("signature")
-                    .and_then(|s| s.as_str())
-                    .map(|s| s.to_string()),
-                content: v
-                    .get("snippet")
-                    .and_then(|s| s.as_str())
-                    .map(|s| s.to_string()),
-                context_prev: None,
-                context_next: None,
-                source: None,
-                chunk_ref: None,
-            })
-            .collect(),
-    }
-}
-
-/// Convert a remote search hit into a local `SearchResultItem`, tagging it with
-/// its origin (`source`) and a project-namespaced `chunk_ref` for later
-/// retrieval.
-///
-/// The `chunk_ref` is `"<peer>/<remote_alias>:<chunk_id>"`. The `remote_alias`
-/// segment is essential: the peer is itself multi-repo and chunk_ids are only
-/// unique *within* a single index, so `federated_get_chunk` must forward the
-/// alias as a `project=` scope to disambiguate. Omitting it (the old
-/// `"<peer>:<id>"` shape) made every remote `get_chunk` fail with
-/// `ambiguous_chunk_id` whenever the peer hosted more than one project.
-///
-/// Literal hits carry no chunk id: both `chunk_id` and `chunk_ref` stay
-/// `None` (the fields are omitted from the rendered JSON). Never substitute
-/// a default here — a fabricated `chunk_id: 0` invites callers to hand-build
-/// a `"<peer>/<alias>:0"` ref that resolves to an unrelated chunk.
-fn convert_remote_item(
-    peer_name: &str,
-    remote_alias: &str,
-    item: crate::federation::RemoteSearchItem,
-) -> SearchResultItem {
-    let chunk_ref = item
-        .chunk_id
-        .map(|id| format!("{peer_name}/{remote_alias}:{id}"));
-    SearchResultItem {
-        chunk_id: item.chunk_id,
-        path: item.path,
-        start_line: item.start_line,
-        end_line: item.end_line,
-        kind: item.kind.unwrap_or_default(),
-        score: item.score,
-        signature: item.signature,
-        content: item.content.or(item.snippet),
-        context_prev: item.context_prev,
-        context_next: item.context_next,
-        source: Some(format!("{peer_name}/{remote_alias}")),
-        chunk_ref,
-    }
-}
-
-/// Apply a `filter_path` prefix filter to federated results **client-side**,
-/// on the namespaced paths the caller actually sees.
-///
-/// Federated `filter_path` cannot be forwarded to the peer: the peer matches
-/// against its own un-namespaced store paths (and, in serve mode, against the
-/// wrong project root), so a server-side match returns nothing for any value.
-/// Here we match against the `<peer>/<alias>/…` path carried on each converted
-/// item, with an empty project root (the namespaced path is already relative),
-/// so the filter means exactly what the caller reads back in the results.
-///
-/// A blank/whitespace filter is a no-op. Returns immediately when `filter_path`
-/// is `None`, so the non-filtered fast path pays nothing.
-fn retain_by_filter_path(items: &mut Vec<SearchResultItem>, filter_path: Option<&str>) {
-    let Some(raw) = filter_path else { return };
-    if raw.trim().is_empty() {
-        return;
-    }
-    let normalized = crate::cache::normalize_filter_path(raw);
-    if normalized.is_empty() {
-        return;
-    }
-    items.retain(|it| crate::cache::path_matches_filter(&it.path, &normalized, ""));
-}
-
-/// True when `filter_path` carries a meaningful prefix (non-blank, non-empty
-/// after normalization) — the single predicate the federated search paths use
-/// to decide whether to over-fetch and post-filter. Mirrors the no-op guards in
-/// [`retain_by_filter_path`] so `has_filter` and the retain stay in lockstep.
-fn is_meaningful_filter(filter_path: Option<&str>) -> bool {
-    filter_path
-        .map(|f| !f.trim().is_empty() && !crate::cache::normalize_filter_path(f).is_empty())
-        .unwrap_or(false)
-}
-
-/// Parse a federated `chunk_ref` into its `(peer, remote_alias, chunk_id)`
-/// parts.
-///
-/// Accepts the current project-namespaced shape `"<peer>/<alias>:<id>"` and,
-/// for backward compatibility, the legacy `"<peer>:<id>"` shape (no alias →
-/// `None`, which falls back to group-scoped lookup on the peer).
-///
-/// The `chunk_id` is taken after the *last* `':'` so peer/alias segments that
-/// themselves contain a colon are not misparsed; the peer/alias split is on the
-/// *first* `'/'`.
-fn parse_federated_chunk_ref(chunk_ref: &str) -> Option<(&str, Option<&str>, u32)> {
-    let (left, id_str) = chunk_ref.rsplit_once(':')?;
-    let chunk_id: u32 = id_str.parse().ok()?;
-    match left.split_once('/') {
-        Some((peer, alias)) if !peer.is_empty() && !alias.is_empty() => {
-            Some((peer, Some(alias), chunk_id))
-        }
-        _ => Some((left, None, chunk_id)),
-    }
-}
-
-/// Best-effort extraction of the concatenated text content of a
-/// `CallToolResult`. Resilient to rmcp's internal content enum shape.
-fn extract_call_tool_text(result: &CallToolResult) -> String {
-    serde_json::to_value(result)
-        .ok()
-        .and_then(|v| {
-            v.get("content").and_then(|c| c.as_array()).map(|arr| {
-                arr.iter()
-                    .filter_map(|item| item.get("text").and_then(|t| t.as_str()))
-                    .collect::<Vec<_>>()
-                    .join("\n")
-            })
-        })
-        .unwrap_or_default()
-}
-
 // ════════════════════════════════════════════════════════════════
 // REST API handlers (federation-friendly HTTP+JSON mirror of MCP tools).
 //
@@ -7487,747 +6743,3 @@ impl ServerHandler for CodesearchService {
             )
     }
 }
-
-// === Server Entry Point ===
-
-/// Run the MCP server using stdio transport with file watching for live index updates.
-///
-/// MCP server mode: how `codesearch mcp` connects to the index backend.
-///
-/// - **Auto** — If `codesearch serve` is running, connect as an HTTP client;
-///   otherwise fall back to local stdio mode.
-/// - **Client** — Always connect to `codesearch serve` via HTTP; fail if not running.
-/// - **Local** — Always use local DB in stdio mode (classic behavior).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, serde::Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum McpMode {
-    /// Connect to serve if available, otherwise local.
-    #[default]
-    Auto,
-    /// Always connect to serve; fail if unreachable.
-    Client,
-    /// Always use local DB (stdio).
-    Local,
-}
-
-impl std::fmt::Display for McpMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            McpMode::Auto => write!(f, "auto"),
-            McpMode::Client => write!(f, "client"),
-            McpMode::Local => write!(f, "local"),
-        }
-    }
-}
-
-impl std::str::FromStr for McpMode {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "auto" => Ok(McpMode::Auto),
-            "client" => Ok(McpMode::Client),
-            "local" => Ok(McpMode::Local),
-            other => Err(format!(
-                "invalid MCP mode '{}': must be 'auto', 'client', or 'local'",
-                other
-            )),
-        }
-    }
-}
-
-/// Probe the serve health endpoint. Returns Ok(serve_url) if serve is alive.
-async fn probe_serve_health(serve_url: &str) -> bool {
-    let client = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_millis(
-            crate::constants::MCP_HEALTH_PROBE_TIMEOUT_MS,
-        ))
-        .build();
-    let Ok(client) = client else { return false };
-    let url = format!("{}{}", serve_url, crate::constants::HEALTH_PATH);
-    client.get(&url).send().await.is_ok()
-}
-
-/// Run `codesearch mcp` as an HTTP client connecting to a running serve instance.
-///
-/// Uses rmcp's `StreamableHttpClientWorker` with `reqwest::Client` to speak
-/// MCP Streamable HTTP to the serve hub. The MCP client (e.g. Claude Code)
-/// talks JSON-RPC over stdio to us, and rmcp relays to the serve HTTP endpoint.
-/// Run `codesearch mcp` as a transparent stdio↔HTTP proxy to `codesearch serve`.
-///
-/// Architecture:
-///   Claude Desktop ──(stdio JSON-RPC)──▶ McpProxyService ──(HTTP Streamable)──▶ codesearch serve
-///
-/// Every MCP request from Claude Desktop is forwarded verbatim to the serve hub and the
-/// response is returned unchanged. This allows Claude Desktop — which has no repo context
-/// of its own — to reach all repos managed by `codesearch serve`.
-///
-/// ## Reconnect behaviour
-///
-/// When `codesearch serve` goes away (restart, crash, network blip), the proxy does NOT
-/// exit. Instead it:
-/// 1. Keeps the stdio connection to Claude Desktop alive
-/// 2. Returns "reconnecting" errors for any incoming tool calls
-/// 3. Retries the HTTP connection every 3 seconds for up to 5 minutes
-/// 4. On success, hot-swaps the peer — tool calls resume immediately
-/// 5. After 5 minutes of failure, exits cleanly (Claude Desktop detects the disconnect)
-///
-/// ## Idle disconnect behaviour
-///
-/// One HTTP MCP session held open for the lifetime of the proxy keeps a request
-/// permanently registered at the remote's ingress, so a scale-to-zero host never
-/// sees 0 concurrent requests and never suspends the replica. To avoid that, the
-/// connection is only held while it is actually being used:
-///
-/// - An idle-checker ticks every `MCP_PROXY_IDLE_CHECK_INTERVAL_SECS`. Once no
-///   request has been forwarded for `CODESEARCH_MCP_PROXY_IDLE_DISCONNECT_SECS`
-///   (default `DEFAULT_MCP_PROXY_IDLE_DISCONNECT_SECS`; `0` disables and restores
-///   the always-connected behaviour), it clears the peer and cancels the
-///   `RunningService`, closing the transport.
-/// - That is a *planned* close, not an outage: it does not open a failure window
-///   and does not count against `reconnect::MAX_DURATION_SECS`. The monitor task's
-///   resulting `disconnect_tx` signal is recognised (via `voluntary_disconnect`)
-///   and does not trigger an eager reconnect — reconnecting immediately would
-///   defeat the purpose.
-/// - The next `list_tools` / `call_tool` finds an empty peer slot and signals
-///   `connect_request_tx`, which reconnects on demand. Failure-path reconnects
-///   are unaffected and still run on their own cadence.
-async fn run_mcp_client(serve_url: &str, cancel_token: CancellationToken) -> Result<()> {
-    use rmcp::{transport::stdio, ServiceExt};
-    use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-
-    let mcp_url = format!("{}{}", serve_url, crate::constants::MCP_ENDPOINT_PATH);
-    tracing::info!("🔗 Connecting to codesearch serve at {}", mcp_url);
-
-    // Channels: spawned monitor tasks notify us when their connection drops.
-    let (disconnect_tx, mut disconnect_rx) = tokio::sync::mpsc::channel::<()>(1);
-    let (stdio_close_tx, mut stdio_close_rx) = tokio::sync::mpsc::channel::<()>(1);
-    // Capacity 1: coalescing duplicate "connect now" requests is correct.
-    let (connect_request_tx, mut connect_request_rx) = tokio::sync::mpsc::channel::<()>(1);
-
-    // Shared peer state — hot-swapped on reconnect.
-    let peer_state: std::sync::Arc<tokio::sync::RwLock<Option<rmcp::service::Peer<RoleClient>>>> =
-        std::sync::Arc::new(tokio::sync::RwLock::new(None));
-
-    // Idle-disconnect state, shared with the proxy service.
-    let last_activity: Arc<Mutex<std::time::Instant>> =
-        Arc::new(Mutex::new(std::time::Instant::now()));
-    let in_flight: Arc<AtomicUsize> = Arc::new(AtomicUsize::new(0));
-    // Notified whenever an on-demand `connect_to_serve` attempt (below, in the
-    // `connect_request_rx` arm) comes back `Err` — lets `await_peer` stop waiting
-    // on a definitive refusal instead of polling out the rest of its window.
-    let connect_failed: Arc<tokio::sync::Notify> = Arc::new(tokio::sync::Notify::new());
-    // Cancellation handle for the *current* connection. `RunningServiceCancellationToken`
-    // is not Clone and its `cancel()` consumes self, so it lives in an Option slot
-    // that the idle-checker `take()`s.
-    let conn_cancel: Arc<
-        tokio::sync::Mutex<Option<rmcp::service::RunningServiceCancellationToken>>,
-    > = Arc::new(tokio::sync::Mutex::new(None));
-    // Set just before we cancel a connection ourselves, so the disconnect signal it
-    // produces is not mistaken for an outage.
-    let voluntary_disconnect = Arc::new(AtomicBool::new(false));
-    let idle_disconnect_secs = resolve_proxy_idle_disconnect_secs(None);
-    if idle_disconnect_secs == 0 {
-        tracing::info!("idle-disconnect disabled — holding the serve connection open");
-    } else {
-        tracing::info!(
-            "💤 idle-disconnect enabled: closing the serve connection after {}s without traffic (checked every {}s)",
-            idle_disconnect_secs,
-            crate::constants::MCP_PROXY_IDLE_CHECK_INTERVAL_SECS
-        );
-    }
-
-    // Step 1: Start stdio proxy for Claude Desktop.
-    // This must happen first so Claude Desktop has something to talk to,
-    // even before the serve connection is established.
-    let proxy = McpProxyService {
-        peer: peer_state.clone(),
-        disconnect_tx: disconnect_tx.clone(),
-        connect_request_tx: connect_request_tx.clone(),
-        last_activity: last_activity.clone(),
-        in_flight: in_flight.clone(),
-        connect_failed: connect_failed.clone(),
-    };
-    let server = proxy
-        .serve(stdio())
-        .await
-        .context("Failed to start proxy stdio server")?;
-
-    // Spawn a task that watches the stdio connection (takes ownership of server).
-    tokio::spawn(async move {
-        let _ = server.waiting().await;
-        let _ = stdio_close_tx.send(()).await;
-    });
-
-    // Step 2: Initial connection to serve (tolerant — may not be running yet).
-    let mut serve_down_since: Option<std::time::Instant> = None;
-    match connect_to_serve(
-        &mcp_url,
-        &peer_state,
-        disconnect_tx.clone(),
-        &conn_cancel,
-        &last_activity,
-    )
-    .await
-    {
-        Ok(()) => {
-            tracing::info!("🚀 MCP proxy ready — forwarding Claude Desktop ↔ codesearch serve");
-        }
-        Err(e) => {
-            serve_down_since = Some(std::time::Instant::now());
-            tracing::warn!(
-                "codesearch serve not yet available ({}). Proxy is up, will retry every {}s.",
-                e,
-                reconnect::INTERVAL_SECS
-            );
-            // Seed a synthetic disconnect so the main loop starts reconnecting.
-            let tx = disconnect_tx.clone();
-            tokio::spawn(async move {
-                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                let _ = tx.send(()).await;
-            });
-        }
-    }
-
-    // Step 3: Main loop — wait for stdio close, serve disconnect, an on-demand
-    // connect request, an idle timeout, or cancel.
-
-    let mut idle_ticker = tokio::time::interval(std::time::Duration::from_secs(
-        crate::constants::MCP_PROXY_IDLE_CHECK_INTERVAL_SECS,
-    ));
-    // The first tick of a tokio interval completes immediately; skip it so a
-    // freshly started proxy is not evaluated for idleness before it can be used.
-    idle_ticker.tick().await;
-
-    loop {
-        tokio::select! {
-            biased; // Prefer clean shutdown paths over reconnect
-
-            // Claude Desktop closed stdio — we're done.
-            _ = stdio_close_rx.recv() => {
-                tracing::info!("MCP proxy transport closed");
-                return Ok(());
-            }
-
-            // External cancel signal (e.g. process termination).
-            _ = cancel_token.cancelled() => {
-                tracing::info!("🛑 Shutdown signal received, stopping MCP proxy...");
-                return Ok(());
-            }
-
-            // A request arrived while the peer slot was empty — connect now rather
-            // than waiting for the failure-path cadence. Ordered before the
-            // disconnect branch so a pending 3s backoff cannot starve it.
-            _ = connect_request_rx.recv() => {
-                if peer_state.read().await.is_some() {
-                    continue; // Someone else already reconnected.
-                }
-                match connect_to_serve(&mcp_url, &peer_state, disconnect_tx.clone(), &conn_cancel, &last_activity).await {
-                    Ok(()) => {
-                        tracing::info!("🔗 Reconnected to codesearch serve on demand");
-                        serve_down_since = None;
-                    }
-                    Err(e) => {
-                        // Serve is genuinely unreachable (or still waking). Hand over
-                        // to the existing failure loop, which retries on its own
-                        // cadence and eventually gives up.
-                        tracing::debug!("On-demand connect failed: {}", e);
-                        note_connect_failure(&connect_failed, &disconnect_tx);
-                    }
-                }
-            }
-
-            // Serve disconnected — enter reconnect loop.
-            _ = disconnect_rx.recv() => {
-                // Clear peer so tool calls get "reconnecting" error.
-                {
-                    let mut p = peer_state.write().await;
-                    *p = None;
-                }
-
-                // A disconnect we caused on purpose (idle-close) is not an outage:
-                // no failure window, no eager reconnect — the next request will ask
-                // for one via connect_request_tx.
-                if voluntary_disconnect.swap(false, Ordering::SeqCst) {
-                    tracing::debug!(
-                        "serve connection closed after idle — will reconnect on the next request"
-                    );
-                    continue;
-                }
-
-                if serve_down_since.is_none() {
-                    serve_down_since = Some(std::time::Instant::now());
-                    tracing::warn!(
-                        "codesearch serve disconnected — will attempt reconnect every {}s for up to {}s",
-                        reconnect::INTERVAL_SECS,
-                        reconnect::MAX_DURATION_SECS,
-                    );
-                }
-
-                let elapsed = serve_down_since.unwrap().elapsed();
-                if elapsed.as_secs() > reconnect::MAX_DURATION_SECS {
-                    tracing::error!(
-                        "❌ Could not reconnect to serve after {}s — giving up",
-                        reconnect::MAX_DURATION_SECS
-                    );
-                    return Ok(()); // Clean exit so Claude Desktop gets graceful EOF
-                }
-
-                // Wait before retrying.
-                tokio::time::sleep(std::time::Duration::from_secs(reconnect::INTERVAL_SECS)).await;
-
-                match connect_to_serve(&mcp_url, &peer_state, disconnect_tx.clone(), &conn_cancel, &last_activity).await {
-                    Ok(()) => {
-                        tracing::info!(
-                            "✅ Reconnected to codesearch serve (was down for {:.0}s)",
-                            serve_down_since.unwrap().elapsed().as_secs()
-                        );
-                        serve_down_since = None;
-                    }
-                    Err(e) => {
-                        tracing::debug!("Reconnect attempt failed: {}", e);
-                        // Re-trigger ourselves: the disconnect_tx from the failed
-                        // connect_to_serve was never used, so we send a synthetic
-                        // disconnect to keep the loop going.
-                        let tx = disconnect_tx.clone();
-                        tokio::spawn(async move {
-                            tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-                            let _ = tx.send(()).await;
-                        });
-                    }
-                }
-            }
-
-            // Idle check — close the connection so a scale-to-zero remote can suspend.
-            _ = idle_ticker.tick() => {
-                if idle_disconnect_secs == 0 {
-                    continue; // Idle-disconnect disabled.
-                }
-                let last = match last_activity.lock() {
-                    Ok(guard) => *guard,
-                    Err(_) => continue, // Poisoned: never tear down on a bookkeeping error.
-                };
-                if !is_idle(last, idle_disconnect_secs, std::time::Instant::now()) {
-                    continue;
-                }
-
-                // Take the peer slot's write lock *before* checking `in_flight`, and
-                // hold it through the clear below, instead of checking `in_flight`
-                // first and taking the write lock afterwards.
-                //
-                // Every forwarding call does `InFlightGuard::new` (increments
-                // `in_flight`) BEFORE `self.peer.read().await` (list_tools/call_tool
-                // above) — so a call that has already obtained `Some(peer)` to
-                // forward through has necessarily already incremented the counter,
-                // and a call that hasn't reached the read yet will block on it once
-                // we hold the write lock. Reading `in_flight` first (the previous
-                // version) left a gap between that read and taking the write lock in
-                // which such a call could still slip past, get `Some(peer)`, and have
-                // its transport cancelled out from under it mid-request — the
-                // in-flight count and the peer-slot teardown were two separate
-                // operations pretending to be one guard. See AGENTS.md
-                // "counter-then-teardown races".
-                let mut p = peer_state.write().await;
-                if p.is_none() {
-                    continue; // Already disconnected.
-                }
-                if in_flight.load(Ordering::SeqCst) > 0 {
-                    continue; // A request is still being forwarded over this transport.
-                }
-
-                tracing::info!(
-                    "💤 Idle for {}s — closing MCP proxy connection to codesearch serve (will reconnect on next request)",
-                    idle_disconnect_secs
-                );
-                // Flag first, so the disconnect signal from the dying monitor task is
-                // recognised as planned no matter how fast it arrives.
-                voluntary_disconnect.store(true, Ordering::SeqCst);
-                *p = None;
-                drop(p);
-                if let Some(token) = conn_cancel.lock().await.take() {
-                    token.cancel();
-                } else {
-                    // Nothing to cancel — don't leave the flag set for a later,
-                    // genuine disconnect to misread.
-                    voluntary_disconnect.store(false, Ordering::SeqCst);
-                }
-            }
-        }
-    }
-}
-
-/// Establish (or re-establish) an HTTP MCP client connection to the serve hub.
-///
-/// On success, updates `peer_state` with the new peer and spawns a background task
-/// that monitors the connection and sends a message on `disconnect_tx` when it drops.
-///
-/// Also parks the connection's cancellation handle in `conn_cancel` (so the
-/// idle-checker can close it) and stamps `last_activity`, so a connection opened
-/// just before an idle tick is not immediately judged idle.
-async fn connect_to_serve(
-    mcp_url: &str,
-    peer_state: &std::sync::Arc<tokio::sync::RwLock<Option<rmcp::service::Peer<RoleClient>>>>,
-    disconnect_tx: tokio::sync::mpsc::Sender<()>,
-    conn_cancel: &Arc<tokio::sync::Mutex<Option<rmcp::service::RunningServiceCancellationToken>>>,
-    last_activity: &Arc<Mutex<std::time::Instant>>,
-) -> Result<()> {
-    use rmcp::ServiceExt;
-
-    let transport = {
-        use rmcp::transport::streamable_http_client::{
-            StreamableHttpClientTransportConfig, StreamableHttpClientWorker,
-        };
-        let config =
-            StreamableHttpClientTransportConfig::with_uri(mcp_url).reinit_on_expired_session(true);
-        StreamableHttpClientWorker::new(reqwest::Client::new(), config)
-    };
-
-    let http_client: rmcp::service::RunningService<RoleClient, ()> =
-        ().serve(transport).await.map_err(|e| {
-            anyhow::anyhow!(
-                "Failed to connect to codesearch serve at {}.\n\
-                 Error: {}\n\
-                 Is `codesearch serve` running?",
-                mcp_url,
-                e
-            )
-        })?;
-
-    // Grab the cancellation handle before the RunningService is moved into the
-    // monitor task below — that's the only way to close this transport later
-    // (idle-disconnect). Cancelling it makes the monitor's `waiting()` resolve, so
-    // the normal disconnect path still runs; nothing needs special-casing.
-    {
-        let mut slot = conn_cancel.lock().await;
-        *slot = Some(http_client.cancellation_token());
-    }
-
-    // Update the shared peer.
-    let peer = http_client.peer().clone();
-    {
-        let mut p = peer_state.write().await;
-        *p = Some(peer);
-    }
-
-    // A fresh connection counts as activity: without this, an idle tick firing
-    // right after a reconnect would immediately close it again.
-    mark_proxy_activity(last_activity);
-
-    // Spawn a monitor task that detects when the connection drops.
-    tokio::spawn(async move {
-        let _ = http_client.waiting().await;
-        // Connection lost — notify main loop.
-        let _ = disconnect_tx.send(()).await;
-    });
-
-    Ok(())
-}
-
-/// # Multi-instance Support
-///
-/// When another instance is already running with write access to the same database,
-/// this server will automatically start in **readonly mode**:
-/// - Searches work normally
-/// - No file watching (index won't auto-update)
-/// - No incremental refresh
-///
-/// This allows multiple terminal windows to use codesearch simultaneously.
-pub async fn run_mcp_server(
-    path: Option<PathBuf>,
-    create_index: bool,
-    log_level: crate::logger::LogLevel,
-    quiet: bool,
-    mode: McpMode,
-    cancel_token: CancellationToken,
-) -> Result<()> {
-    let serve_url = serve_url_from_env();
-
-    // Set FASTEMBED_CACHE_DIR early (before any embedding work) to ensure fastembed
-    // downloads and caches models to ~/.codesearch/models instead of creating
-    // .fastembed_cache in the current working directory. Do this once for all modes.
-    match crate::constants::get_global_models_cache_dir() {
-        Ok(models_dir) => {
-            std::env::set_var("FASTEMBED_CACHE_DIR", &models_dir);
-        }
-        Err(e) => {
-            tracing::warn!("Could not set FASTEMBED_CACHE_DIR: {}", e);
-        }
-    }
-
-    match mode {
-        McpMode::Client => {
-            // Client mode: init logger using global cache dir (no local DB needed)
-            if let Err(e) = crate::logger::init_logger(
-                &crate::constants::get_global_cache_dir(),
-                log_level,
-                quiet,
-            ) {
-                tracing::warn!("Failed to initialize file logger: {}", e);
-            }
-            tracing::info!("📡 MCP mode: client — connecting to serve at {}", serve_url);
-            if !probe_serve_health(&serve_url).await {
-                return Err(anyhow::anyhow!(
-                    "codesearch serve is not running at {}. \
-                     Start it with `codesearch serve` or use --mode auto/local.",
-                    serve_url
-                ));
-            }
-            return run_mcp_client(&serve_url, cancel_token).await;
-        }
-        McpMode::Auto => {
-            // Auto mode: init logger early for probe logging
-            if let Err(e) = crate::logger::init_logger(
-                &crate::constants::get_global_cache_dir(),
-                log_level,
-                quiet,
-            ) {
-                tracing::warn!("Failed to initialize file logger: {}", e);
-            }
-            if probe_serve_health(&serve_url).await {
-                tracing::info!(
-                    "📡 MCP mode: auto — serve detected at {}, connecting as client",
-                    serve_url
-                );
-                return run_mcp_client(&serve_url, cancel_token).await;
-            }
-            tracing::info!("📡 MCP mode: auto — no serve detected, falling back to local stdio");
-            // Fall through to local mode
-        }
-        McpMode::Local => {
-            tracing::info!("📡 MCP mode: local — using local DB (stdio)");
-            // Fall through to local mode
-        }
-    }
-
-    // ── Local stdio mode (original behavior) ──────────────────────────
-    use rmcp::{transport::stdio, ServiceExt};
-
-    tracing::info!("🚀 Starting codesearch MCP server");
-
-    // Use database discovery to find the best database
-    let db_info = find_best_database(path.as_deref())?;
-
-    let (project_path, db_path) = if let Some(info) = db_info {
-        (info.project_path, info.db_path)
-    } else {
-        // No database found
-        if !create_index {
-            return Err(anyhow::anyhow!(
-                "No database found in current directory, parent directories, or globally tracked repositories. \
-                 Run 'codesearch index' first to index the codebase, or use --create-index=true flag to automatically create it."
-            ));
-        }
-
-        // Create minimal database structure to allow server to start immediately
-        let effective_path = path.as_ref().cloned().unwrap_or(std::env::current_dir()?);
-
-        // Use git root detection to place database in the correct location
-        let db_root =
-            crate::index::find_git_root(&effective_path)?.unwrap_or_else(|| effective_path.clone());
-        let db_path = db_root.join(".codesearch.db");
-
-        tracing::info!(
-            "📁 Creating minimal database structure at {}",
-            db_path.display()
-        );
-
-        // Create directory
-        std::fs::create_dir_all(&db_path)?;
-
-        // Get model info
-        let model_type = ModelType::default();
-        let model_short_name = model_type.short_name().to_string();
-        let dimensions = model_type.dimensions();
-
-        // Create minimal metadata.json (atomic read-modify-write, matching format
-        // used by build_index). Routes through the single-source-of-truth stamp so
-        // model_name matches the other index paths (previously wrote the Debug
-        // variant name here, e.g. "AllMiniLML6V2Q", instead of the model name).
-        crate::vectordb::merge_metadata_atomic(&db_path, |obj| {
-            model_type.write_metadata_fields(obj);
-            obj.insert(
-                "indexed_at".to_string(),
-                serde_json::Value::String(chrono::Utc::now().to_rfc3339()),
-            );
-        })?;
-
-        // Create minimal file_meta.json (matching FileMetaStore format)
-        let file_meta = crate::cache::FileMetaStore::new(model_short_name.clone(), dimensions);
-        file_meta.save(&db_path)?;
-
-        // Create FTS directory
-        let fts_path = db_path.join("fts");
-        std::fs::create_dir_all(&fts_path)?;
-
-        // Create LMDB file by opening VectorStore (creates minimal structure)
-        let _store = crate::vectordb::VectorStore::new(&db_path, dimensions)?;
-
-        tracing::info!("✅ Minimal database created successfully");
-        tracing::info!("🔄 Background indexing will begin shortly via incremental refresh");
-
-        (effective_path, db_path)
-    };
-
-    // Initialize file logger now that db_path is known (works for both existing and auto-created DB)
-    // NOTE: For MCP, tracing is NOT initialized in main.rs — this is the only init call
-    if let Err(e) = crate::logger::init_logger(&db_path, log_level, quiet) {
-        tracing::warn!("Failed to initialize file logger: {}", e);
-    }
-
-    tracing::info!("📂 Project: {}", project_path.display());
-    tracing::info!("💾 Database: {}", db_path.display());
-
-    // Read model metadata to get dimensions (fallback to 384 if missing/corrupt)
-    let metadata_path = db_path.join("metadata.json");
-    let dimensions = if metadata_path.exists() {
-        match std::fs::read_to_string(&metadata_path)
-            .ok()
-            .and_then(|c| serde_json::from_str::<serde_json::Value>(&c).ok())
-            .and_then(|j| j.get("dimensions").and_then(|v| v.as_u64()))
-        {
-            Some(d) => d as usize,
-            None => {
-                tracing::warn!(
-                    "⚠️  Could not parse dimensions from metadata.json, using default {}",
-                    crate::constants::DEFAULT_EMBEDDING_DIMENSIONS
-                );
-                crate::constants::DEFAULT_EMBEDDING_DIMENSIONS
-            }
-        }
-    } else {
-        tracing::warn!(
-            "⚠️  metadata.json not found, using default dimensions {}",
-            crate::constants::DEFAULT_EMBEDDING_DIMENSIONS
-        );
-        crate::constants::DEFAULT_EMBEDDING_DIMENSIONS
-    };
-
-    // Create shared stores - try write mode first, fall back to readonly if locked
-    // This enables multiple terminal windows to use the same database
-    tracing::info!("📦 Creating shared stores...");
-    let (shared_stores, is_readonly) = SharedStores::new_or_readonly(&db_path, dimensions)?;
-    let shared_stores = Arc::new(shared_stores);
-
-    if is_readonly {
-        tracing::warn!("🔒 Running in READONLY mode (another instance has write access)");
-        tracing::warn!("   ↳ Searches work normally, but index won't auto-update");
-        tracing::warn!("   ↳ Close the other instance to enable write mode");
-    }
-
-    // Create MCP service with shared stores (ready immediately)
-    let service = CodesearchService::new_with_stores(
-        Some(project_path.clone()),
-        Some(shared_stores.clone()),
-    )?;
-
-    tracing::info!("🧠 Model: {}", service.model_type.name());
-
-    // START MCP SERVER NOW - fixes timeout!
-    tracing::info!(
-        "🚀 Starting MCP server{}...",
-        if is_readonly { " (readonly)" } else { "" }
-    );
-    let server = service.serve(stdio()).await?;
-
-    tracing::info!("MCP server ready. Waiting for requests...");
-
-    // Only run background tasks if we have write access
-    if !is_readonly {
-        // Create IndexManager with shared stores (skip initial refresh - do in background)
-        tracing::info!("🔍 Initializing index manager...");
-        let index_manager =
-            IndexManager::new_without_refresh(&project_path, shared_stores.clone()).await?;
-
-        // Background: refresh FIRST, then file watcher (sequential, not concurrent)
-        // Both write to SharedStores, so they must not run concurrently
-        let project_path_clone = project_path.clone();
-        let db_path_clone = db_path.clone();
-        let shared_stores_clone = shared_stores.clone();
-        let index_manager_arc = Arc::new(index_manager);
-        let bg_cancel_token = cancel_token.clone();
-        tokio::spawn(async move {
-            // Step 0: Pre-start FSW to collect file change events during refresh
-            // This ensures changes made while the refresh is running are not missed
-            if let Err(e) = index_manager_arc.start_watching().await {
-                tracing::warn!("⚠️ Could not pre-start file watcher: {}", e);
-            }
-
-            // Step 1: Run initial refresh (writes to stores)
-            tracing::info!("🔄 Starting background incremental refresh...");
-            match IndexManager::perform_incremental_refresh_with_stores(
-                &project_path_clone,
-                &db_path_clone,
-                &shared_stores_clone,
-                &bg_cancel_token,
-            )
-            .await
-            {
-                Ok(_) => {
-                    tracing::info!("✅ Background incremental refresh completed");
-
-                    // Check if shutdown was requested during refresh
-                    if bg_cancel_token.is_cancelled() {
-                        tracing::info!("🛑 Shutdown requested, skipping file watcher startup");
-                        return;
-                    }
-
-                    // Step 2: AFTER refresh completes, start file watcher (also writes to stores)
-                    tracing::info!("👀 Starting file watcher...");
-                    if let Err(e) = index_manager_arc
-                        .start_file_watcher(bg_cancel_token, None, None)
-                        .await
-                    {
-                        tracing::error!("❌ Failed to start file watcher: {}", e);
-                    } else {
-                        tracing::info!(
-                            "✅ File watcher active - index will auto-update on file changes"
-                        );
-                    }
-                }
-                Err(e) => {
-                    tracing::error!("❌ Background incremental refresh failed: {}", e);
-                }
-            }
-        });
-
-        // Start periodic log cleanup task
-        let db_path_for_cleanup = db_path.clone();
-        let cleanup_cancel_token = cancel_token.clone();
-        tokio::spawn(async move {
-            use crate::logger::{cleanup_old_logs, LogRotationConfig};
-
-            // Run initial cleanup on startup
-            let rotation_config = LogRotationConfig::from_env();
-            tracing::info!("🧹 Running initial log cleanup...");
-            if let Err(e) = cleanup_old_logs(&db_path_for_cleanup, &rotation_config) {
-                tracing::warn!("Initial log cleanup failed: {}", e);
-            }
-
-            // Start periodic cleanup task (every 24 hours by default)
-            crate::logger::start_cleanup_task(
-                db_path_for_cleanup.clone(),
-                rotation_config,
-                cleanup_cancel_token,
-            );
-        });
-    } else {
-        tracing::info!("📖 Readonly mode: skipping background refresh and file watcher");
-    }
-
-    // Wait for shutdown: either MCP transport closes or cancellation token fires
-    tokio::select! {
-        result = server.waiting() => {
-            tracing::info!("MCP server transport closed");
-            result?;
-        }
-        _ = cancel_token.cancelled() => {
-            tracing::info!("🛑 Shutdown signal received, stopping MCP server...");
-        }
-    }
-
-    tracing::info!("✅ MCP server shut down cleanly");
-    Ok(())
-}
-
-#[cfg(test)]
-#[path = "federation_helpers_tests.rs"]
-mod federation_helpers_tests;
